+++ gdsfactory/get_netlist.py
@@ -96,7 +96,8 @@ def get_netlist(
 
         settings = c.get_settings(full_settings=full_settings)
         instances[reference_name] = dict(
-            component=c.function_name,
+            # component=c.function_name,
+            component=getattr(c, "function_name", None) or getattr(c, "name"),
             settings=settings["settings"],
         )
         placements[reference_name] = dict(